### 2018-03-15 #627 @whedon
[PRE REVIEW]: EmissV: An R package to create vehicular and other emissions by Top-down methods to air quality models
**Submitting author:** @Schuch666 (<a href="http://orcid.org/0000-0001-5977-4519">Daniel Schuch</a>)
**Repository:** <a href="https://github.com/atmoschem/EmissV" target ="_blank">https://github.com/atmoschem/EmissV</a>
**Version:** 0.664.3
