### 2018-03-24 d8100bb
for CRAN
