+++ R/lineSource.R
@@ -16,21 +16,28 @@
 #' @seealso \code{\link{gridInfo}} and \code{\link{rasterSource}}
 #'
 #'
-#' @examples \dontrun{
+#' @examples
 #' # Do not run
+#' library("osmar")
+#' roads <- osmar::get_osm(complete_file(),
+#'                           source = osmsource_file(paste(system.file("extdata",package="EmissV"),
+#'                                                "/streets.osm.xz",sep="")))
+#' roads <- osmar::as_sp(roads,what = "lines")
 #'
-#' roads <- readRDS(paste(system.file("extdata", package = "EmissV"),"/streets.rds",sep=""))
 #' d3    <- gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d03"))
 #'
-#' roadLength <- lineSource(roads,d3,as_raster=T)
-#'
+#' roadLength <- lineSource(roads,d3,as_raster=TRUE)
+#' \dontrun{
 #' sp::spplot(roadLength, scales = list(draw=TRUE), ylab="Lat", xlab="Lon",main="Length of roads",
 #'            sp.layout=list("sp.lines", roads))
 #' }
 #'
+#'@source OpenstreetMap data avaliable \url{https://www.openstreetmap.org/}
+#'
 
 # algorithm source:
 # https://gis.stackexchange.com/questions/119993/convert-line-shapefile-to-raster-value-total-length-of-lines-within-cell
+
 
 lineSource <- function(s,grid,as_raster = F,verbose = T){
 
