+++ R/totalvoc.R
@@ -19,13 +19,13 @@
 #'
 #'@import units
 #'
-#'@examples \dontrun{
+#'@examples
 #' # Do not run
 #'
-#' veiculos <- vehicles(exemples = T)
+#' veiculos <- vehicles(example = TRUE)
 #'
 #' voc = c("eth","hc3","hc5","hc8","ol2","olt","oli","iso","tol","xyl","ket","ch3oh","ald")
-#' EF_voc <- as.data.frame(matrix(NA,ncol = 9,nrow = 8,byrow = T),
+#' EF_voc <- as.data.frame(matrix(NA,ncol = 9,nrow = 8,byrow = TRUE),
 #'                         row.names = row.names(veiculos))
 #' names(EF_voc) <-  c("voc_vap_g","voc_vap_e","voc_vap_d",
 #'                     "voc_liq_g","voc_liq_e","voc_liq_d",
@@ -44,7 +44,6 @@
 #' print(EF_voc)
 #'
 #' VOC <- totalVOC(veiculos,EF_voc,pol = voc[10])
-#'}
 
 totalVOC <- function(v,ef,pol,verbose=T){
 
