### 2018-03-26 #627 @arfon
@Schuch666 - could you take a look at the reviewer list (https://bit.ly/joss-reviewers) and see if you could suggest a few potential reviewers?
