### 2018-03-26 #627 @arfon
> Hi @arfon , is it possible to suggest you as editor?

:-) Sure thing. I'd like to find a couple of reviewers first.
