### 2018-03-28 #627 @arfon
@benmarwick @nuest would you be willing to review this submission for JOSS?
