+++ R/plumerise.R
@@ -5,8 +5,9 @@
 #' @format numeric value in (units) meters
 #'
 #' @param df data.frame with micrometeorological and source data
-#' @param imax number of maximum interations
-#' @param errmax maximum error
+#' @param imax maximum number of iteractions
+#' @param ermax maximum error
+#' @param Hmax use weil limit for plume rise, see details
 #' @param verbose display additional information
 #'
 #' @export
@@ -14,71 +15,93 @@
 #' @examples \dontrun{
 #' # Do not run
 #'
-#' chimney <- matrix(c(1,1,1,1,1,1,1,1,1,1),ncol = 10, byrow = T)
+#' chimney <- matrix(c(20,0.5,12,400,4,280,100,1,3.6,2.8),ncol = 10, byrow = T)
 #' chimney<- data.frame(chimney)
-#' names(chimney) <- c("z","r","Ve","Te","h","ws","Temp","Ustar","Wstar","L")
+#' names(chimney) <- c("z","r","Ve","Te","ws","Temp","h","L","Ustar","Wstar")
 #'
 #' chimay <- plumeRise(chimney)
 #'
 #'}
 
-plumeRise <- function(df, imax = 100, errmax = 0.1, verbose = T)
+plumeRise <- function(df, imax = 1000, ermax = 1/100, Hmax = T, verbose = T)
 {
-  g    <- 9.81      # m / s2
+  g      <- 9.81      # m / s2
+  rise   <- rep(0,nrow(df))
+  He     <- rep(0,nrow(df))
 
-  # source parameters
-  Hs   <- df$z      # altura da fonte
-  r    <- df$r      # raio da fonte
-  Vi   <- df$ve     # velocidade de exaustão
-  Ti   <- df$Te     # temperatura de exaustão
+  for(j in 1:nrow(df)){
+    # source parameters
+    Hs   <- df$z[j]      # altura da fonte
+    r    <- df$r[j]      # raio da fonte
+    Vi   <- df$Ve[j]     # velocidade de exaustão
+    Ti   <- df$Te[j]     # temperatura de exaustão
 
-  # micrometeorological data
-  h    <- df$h      # altura da camada limite planetária
-  L    <- abs(df$L) # comp monin-obukov
-  U    <- df$ws     # vel média do vento
-  Wstar<- df$Wstar  # w*
-  Ustar<- df$Ustar  # u*
-  Ta   <- df$Temp   # temperatura
+    # micrometeorological data
+    U    <- df$ws[j]     # vel média do vento
+    Ta   <- df$Temp[j]   # temperatura
+    h    <- df$h[j]      # altura da camada limite planetária
+    L    <- abs(df$L[j]) # comp monin-obukov
+    Wstar<- df$Wstar[j]  # w*
+    Ustar<- df$Ustar[j]  # u*
 
-  # termo de flutuabilidade
-  Flu  <- g * Vi * r^2 * (Ti - Ta)/Ti
+    # termo de flutuabilidade
+    Flu  <- g * Vi * r^2 * (Ti - Ta)/Ti
 
-  # convecção forte
-  if(h/L > 10){
-    deltaH <- 4.3 * (Flu / (U * Wstar^2))^(3/5) * h^(2/5)
-  }else
-  # moderadamente convectivas
-  if(h/L <= 10){
-    Wd     <- 0.4 * Wstar # downdrafts mean speed
-    deltaH <- Flu / U*Wd^2
-    for(i in 1:imax){
-      old <- deltaH
-      deltaH <- ( Flu / U*Wd^2 )^(3/5) * (1 + 2*Hs/deltaH)^2
-      err    <- (deltaH - old )/old
-      if(err <= errmax)
-        i = imax
+    # convecção forte
+    if(h/L > 10){
+      if(verbose)
[…]
