+++ R/plumerise.R
@@ -58,15 +58,25 @@ plumeRise <- function(df, imax = 1000, ermax = 1/100, Hmax = T, verbose = T)
         if(verbose)
           print(paste("convective, h/L =",h/L))
         Wd     <- 0.4 * Wstar # downdrafts mean speed
-        deltaH <- Flu / U*Wd^2
+        a      <- ( Flu / U*Wd^2 )^(3/5)
+        b      <- 2 * Hs
+        deltaH <- b
         for(i in 1:imax){
-          old <- deltaH
-          deltaH <- ( Flu / U*Wd^2 )^(3/5) * (1 + 2*Hs/deltaH)^2
-          err    <- (deltaH - old )/old
-          if(err <= ermax){
-            # print(paste(i,"iterações"))
-            i = imax
-          }
+          old    <- deltaH
+          deltaH <- a * (1 + b/old)^2
+          # deltaH <- old + 0.1 * (old^3 / 2* b) * ((1 + b/old)^2 / (2*b*(b + old)))
+          # deltaH<- old - 0.1 * (3*b + old)/(2 * old)
+          err    <- abs( (deltaH - old )/old )
+          print(paste("OLD=",old))
+          print(paste("Dh=",deltaH))
+          print(paste("err=",err))
+          print("========================================")
+          # print(paste("Dh=",deltaH))
+          # if(err <= ermax){
+          #   print(paste(i,"iterações"))
+          #   print(paste("Dh=",deltaH))
+          #   i = imax
+          # }
         }
       }else
         # neutral
@@ -105,3 +115,90 @@ plumeRise <- function(df, imax = 1000, ermax = 1/100, Hmax = T, verbose = T)
   df <- cbind(df,He)
   return(df)
 }
+# !     ******************* PLUME RISE **********************
+#
+#   !      SUBROUTINE PRISE(L,Q,QEFF,ALT,WE,UR,AFOE,HF,AFO,DH,US,NE,NR,FBE,FB,ES,pkr,tss,tff,d,vs)
+# SUBROUTINE PRISE(L,Q,QEFF,ALT,WE,UR,AFOE,AFO,DH,US,FBE,FB,ES,tss,tff,d,vs)
+# IMPLICIT REAL*8 (A-Z)
+# REAL*8 USTAR,ELLE,WSTAR,HMX,ES,tff,tss,d,vs	  !hf,
+# REAL*8 HS,US,TS,AFO,FB,GRADT,DH,ESSE,QEFF,FBE !,AA,AJ,AL,GU,PU
+# !INTEGER NE,NR,pkr
+#
+# d=9.
+# FB=9.81*((D/2.)**2.)*VS*((Tss-Tff)/Tss)
+# FBE=FB/(US*ALT*(WE**2.))
+# !      FBE=FB/(US*ALT*(WE**2.))
+# USTAR=UR
+# WSTAR=WE
+# HMX=ALT
+# TS=300.
+# !	ts=tff  !ts(temperatura ambiente),gradt(grad de temperatura)
+# GRADT=0.01
+# HS=AFO
+# ELLE=L
+# DHCONV1=10000.
+# DHCONV2=10000.
+# DHSTAB1=10000.
+# DHSTAB2=10000.
[…]
