+++ R/plumerise.R
@@ -1,10 +1,10 @@
 #' Calculate plume rise from Brigs (1975) formulation.
 #'
-#' @description Return plume rise height
+#' @description Calculate the maximum height of rise based on Brigs (1975), the height is calculated using different formulations depending on stability and wind conditions.
 #'
-#' @format numeric value in (units) meters
+#' @format data.frame with the input, rise (m) and effective higt (m)
 #'
-#' @param df data.frame with micrometeorological and source data
+#' @param df data.frame with micrometeorological and emission data
 #' @param imax maximum number of iteractions
 #' @param ermax maximum error
 #' @param Hmax use weil limit for plume rise, see details
@@ -12,19 +12,60 @@
 #'
 #' @export
 #'
-#' @examples \dontrun{
+#' @references The plume rise formulas are from Brigs (1975):"Brigs, G. A. Plume rise predictions, Lectures on Air Pollution and Environmental Impact Analyses. Amer. Meteor. Soc. p. 59-111, 1975." and Arya 1999: "Arya, S.P., 1999, Air Pollution Meteorology and Dispersion, Oxford University Press, New York, 310 p."
+#'
+#' The limits are from Weil (1979): "WEIL, J.C. Assessmet of plume rise and dispersion models using LIDAR data, PPSP-MP-24. Prepared by Environmental Center, Martin Marietta Corporation, for Maryland Department of Natural Resources. 1979."
+#'
+#' The example is data from a chimney of the Candiota thermoelectric powerplant from Arabage et al (2006) "Arabage, M. C.; Degrazia, G. A.; Moraes O. L. Simulação euleriana da dispersão local da pluma de poluente atmosférico de Candiota-RS. Revista Brasileira de Meteorologia, v.21, n.2, p. 153-160, 2006."
+#'
+#' @details The input data.frame must contains the folloging colluns:
+#'
+#' - z: height of the emission (m)
+#'
+#' - r: source raius (m)
+#'
+#' - Ve: emission velocity (m/s)
+#'
+#' - Te: emission temperature (K)
+#'
+#' - ws: wind speed (m/s)
+#'
+#' - Temp: ambient temperature (K)
+#'
+#' - h: height of the Atmospheric Boundary Layer-ABL (m)
+#'
+#' - L: Monin-Obuhkov Lench (m)
+#'
+#' - dtdz: lapse ration of potential temperature, used only for stable ABL (K/m)
+#'
+#' - Ustar: atriction velocity, used only for neutral ABL (m/s)
+#'
+#' - Wstar: scale of convectie velocity, used only for convective ABL (m/s)
+#'
+#' Addcitionaly some combination of ws, Wstar and Ustar can produce inacurate results, Weil (1979) propose a geometric limit of 0.62 * (h - Hs) for the rise value.
+#'
+#' @examples
 #' # Do not run
 #'
-#' chimney <- matrix(c(20,0.5,12,400,4,280,100,1,3.6,2.8),ncol = 10, byrow = T)
-#' chimney<- data.frame(chimney)
[…]
