+++ R/rasterSource.R
@@ -22,7 +22,7 @@
 #' x     <- raster::raster(paste(system.file("extdata", package = "EmissV"),"/dmsp.tiff",sep=""))
 #' test  <- rasterSource(x,grid)
 #' \dontrun{
-#' image(test, main = "Spatial distribution by Persistent Nocturnal Lights from DMSP")
+#' image(test,axe = F, main = "Spatial distribution by Persistent Nocturnal Lights from DMSP")
 #'}
 #'
 #'@source Data avaliable \url{https://ngdc.noaa.gov/eog/dmsp/downloadV4composites.html}
