+++ man/pointSource.Rd
@@ -33,8 +33,8 @@ p = data.frame(lat      = c(-22,-22,-23.5),
 
 p_emissions <- pointSource(emissions = p, grid = d1)
 \dontrun{
-spplot(p_emissions,scales = list(draw=TRUE), ylab="Lat", xlab="Lon",
-       main = "3 point sources for domain d1")
+sp::spplot(p_emissions,scales = list(draw=TRUE), ylab="Lat", xlab="Lon",
+           main = "3 point sources for domain d1")
 }
 
 }

[…]
