+++ R/rasterSource.R
@@ -29,7 +29,7 @@
 #'@details About the DMSP and example data \url{https://en.wikipedia.org/wiki/Defense_Meteorological_Satellite_Program}
 
 
-rasterSource <- function(r,grid,verbose = T){
+rasterSource <- function(r,grid,nlevels="all",verbose = T){
 
   col   <- grid$Horizontal[1]
   rol   <- grid$Horizontal[2]
@@ -39,16 +39,36 @@ rasterSource <- function(r,grid,verbose = T){
                           xmn=r.lon[1],xmx=r.lon[2],ymn=r.lat[1],ymx=r.lat[2],
                           crs=sp::proj4string(r))
 
-  total_box <- cellStats(raster::crop(r,box),"sum",na.rm=TRUE)
+  if(is.na(grid$z[1])){
+    total_box <- cellStats(raster::crop(r,box),"sum",na.rm=TRUE)
 
-  X    <- raster::resample(r,box,method = "bilinear") # non-conservative transformation
-  X    <- raster::flip(X,2)
-  X    <- raster::t(X)
-  X    <- raster::as.matrix(X)
-  X    <- X * total_box/sum(X) # to conserve mass
+    X    <- raster::resample(r,box,method = "bilinear") # non-conservative transformation
+    X    <- raster::flip(X,2)
+    X    <- raster::t(X)
+    X    <- raster::as.matrix(X)
+    X    <- X * total_box/sum(X) # to conserve mass
 
-  if(verbose)
-    print(paste("Grid output:",col,"columns",rol,"rows"))
+    if(verbose)
+      print(paste("Grid output:",col,"columns",rol,"rows"))
+  }else{
+    if(nlevels == "all"){
+      nlevels <- dim(d1$z)[3]
+    }else{
+      nlevels <- nlevels
+    }
+    total_box <- cellStats(raster::crop(r,box),"sum",na.rm=TRUE)
+
+    X    <- raster::resample(r,box,method = "bilinear") # non-conservative transformation
+    X    <- raster::flip(X,2)
+    X    <- raster::t(X)
+    X    <- raster::as.array(X)
+    X    <- X[,,1:nlevels]
+    X    <- X * total_box[1:nlevels]/sum(X) # to conserve mass
+
+    if(verbose)
+      print(paste("Grid output:",col,"columns",rol,"rows",nlevels,"levels"))
+  }
+
   return(X)
 }
 
