### 2018-05-23 atmoschem/EmissV#11 @nuest
License and author documentation
(part of https://github.com/openjournals/joss-reviews/issues/662)

The requirements for the software license are fulfilled, however as a reader I still find it useful to have a reference to the license and the authors in the README file. I suggest to add a short section to the end of the README.
