### 2018-05-23 atmoschem/EmissV#12 @nuest
Restructure and extend README
(part of https://github.com/openjournals/joss-reviews/issues/662)

I suggest to restructure the README file to better align with the required sections:

```
# EmissV

statement of need, in plain language

## Installation

### System dependencies

Ubuntu, Fedora, ...

### Package installation

devtools::...

## Using `EmissV`

[R code chunk]

## ...

[add info about contribution, code of conduct, reporting bugs, seeking support, ...]

## License

...

```

- Can you provide installation instructions for Windows? The Appveyor build works, so I assume a Windows installation is possible. Possibly just mention that no specific steps are needed.
- Regarding the "needed packages"
  - This list will at some point be out of sync with your `DESCRIPTION` file, which is "Good" according to the reviewer guidelines already. I suggest to only list packages here that you actually provide information about, like how/why you import them, what they are needed for - and only if it is not obvious (like sp and raster for geospatial data)
  - "The dependency of the packages spatsts, maptools and sp will be removed son." This is potentially confusing if you do not provide motivation. I suggest to remove this from README and instead create an issue stating that these packages should be removed _any why_.
- For the Example, I suggest you switch to using a `README.Rmd` file to generate the README.md with a R code based example, see http://usethis.r-lib.org/reference/use_readme_rmd.html
- Meta-Information related to _Community guidelines_ are missing, check the ROpenSci GitHub org for examples (such as https://github.com/ropensci/nomisr#meta)
