### 2018-05-23 atmoschem/EmissV#11 @nuest
You can also use that section in the README to clarify the involvement of the third author listed in `paper.md` (
Edmilson Freitas ) who did not contribute via GitHub (https://github.com/atmoschem/EmissV/graphs/contributors).
