### 2018-05-23 atmoschem/EmissV#13 @nuest
Add automated tests
(part of https://github.com/openjournals/joss-reviews/issues/662)

Strongly recommend to use `testthat` to add tests for the package functionality.
AFAICS the Travis and Appveyor tests only run an installation of the package, but no unit or integration tests. Pointers welcome if I'm wrong!

It is also worth having examples in the docs that actually run (I saw mostly "dontrun" code blocks).
