### 2018-05-23 atmoschem/EmissV#14 @nuest
Re-check language of examples and comments
(part of https://github.com/openjournals/joss-reviews/issues/662)

It's fun to learn some Portuguese and always refreshing to interact with different languages and cultures. However, for the code examples, I suggest to stick to English terms for clarity.

- veiculos: https://github.com/atmoschem/EmissV/search?utf8=%E2%9C%93&q=veiculos

Translating some source code comments might be more welcoming to non-Portuguese contributors, e.g. https://github.com/atmoschem/EmissV/blob/master/R/plumerise.R
