### 2018-05-23 #662 @nuest
Sincere apologies for the delay, too many GH emails and this one fell through the cracks. I started my review and have some first comments that can be addressed rightaway, see issues linked above.
I'll start the functional evaluation next.
