+++ DESCRIPTION
@@ -5,7 +5,7 @@ Authors@R: c(
   person(given = "Daniel",
          family = "Schuch",
          role = c("aut", "cre"),
-         email = "underschuch@gmail.com",
+         email = "schuch@usp.br",
          comment = c(ORCID = "0000-0001-5977-4519")),
   person(given = "Sergio",
          family = "Ibarra-Espinosa",
