+++ DESCRIPTION
@@ -5,7 +5,7 @@ Authors@R: c(
   person(given = "Daniel",
          family = "Schuch",
          role = c("aut", "cre"),
-         email = "underschuch@gmail.com",
+         email = "schuch@usp.br",
          comment = c(ORCID = "0000-0001-5977-4519")),
   person(given = "Sergio",
          family = "Ibarra-Espinosa",
@@ -13,7 +13,7 @@ Authors@R: c(
          email = "sergio.ibarra@usp.br",
          comment = c(ORCID = "0000-0002-3162-1905"))
   )
-Description: Top-down methods to create vehicular and other emissions.
+Description: Top-down methods to create vehicular and other emissions for atmospheric air quality models.
 Depends: R (>= 3.4.1)
 Imports: ncdf4, raster, units, methods, sf, data.table, sp, spatstat, maptools, rgeos, rgdal
 Suggests: eixport, osmar
