### 2018-05-23 atmoschem/EmissV#12 @nuest
I can confirm the system dependency list you have in the README, tested it with Docker using `rocker/r-ver:3.5.0` for Ubuntu and `rhub/fedora-gcc-devel` for Fedora. I was able to install the package when following your installation instructions with [this additional trick](https://gist.github.com/giocomai/0a94e49b3dad053292e73fce7feda411) for `udunits`. Anyway, good job! :100:

Installation on Windows also worked without problems by the way, thanks to the people working hard on Windows binaries.

_Minor comment_: Why do you break up the `yum install` into two commands? I assume it is because the first one is for sf _and_ units and the second one is for netcdf... I recommend following the same structure as for Ubuntu and split the installation up per R package, and add the same comments to Fedora. Right now Fedora users have to make sense of the Ubuntu instructions to understand why they should run the commands.

One further thought:

_What do you think about referencing the installation instructions of `udunits` and `sf` (i.e. https://r-spatial.github.io/sf/#installing) instead of adding them yourself?_ IMHO this has the advantage of not having outdated instructions.
