### 2018-05-23 atmoschem/EmissV#15 @nuest
Clarify documentation
(part of https://github.com/openjournals/joss-reviews/issues/662)

- `?vehicles`
  - tells me it returns "data frame with lines by vehicle category and columns for category, type, Fuel, use and a additional column for each area." - **What does the value in the column for each area mean?**
  - "diary use" - do you mean "average daily kilometres driven by the vehicle type" ? Recommend to be a bit more verbose here.
