### 2018-05-23 #662 @benmarwick
Yes, to start with, I'd love to see another attempt at a clear statement of what problems the software is designed to solve and who the target audience is? Who are the users like to be? I've just done a tiny bit work helping students in atmospheric chemistry, so the methods here are not familiar to me.

I can run the example code until here:

```
TOTAL  <- totalEmission(veiculos,EF,pol = c("CO"),verbose = T)
Error: $ operator is invalid for atomic vectors
5.
as.character.units(structure(1127549.32468524, units = structure(list(
    numerator = "t", denominator = "year"), .Names = c("numerator",
"denominator"), class = "symbolic_units"), class = "units"))
4.
as.character(structure(1127549.32468524, units = structure(list(
    numerator = "t", denominator = "year"), .Names = c("numerator",
"denominator"), class = "symbolic_units"), class = "units"))
3.
paste("Total of", pol[i], ":", sum(total_t_y), units::deparse_unit(total_t_y))
2.
print(paste("Total of", pol[i], ":", sum(total_t_y), units::deparse_unit(total_t_y))) at totalEmission.R#62
1.
totalEmission(veiculos, EF, pol = c("CO"), verbose = T)
```
