### 2018-05-23 32b8025
read-me update for JOSS
