### 2018-05-23 atmoschem/EmissV#11 @Schuch666
Hi @nuest
I added the short description of the license and the package
