### 2018-05-23 #662 @Schuch666
Thank @nuest and @benmarwick

I'm working on the text (and the code). Soon I will update the package and write here some detailed explanations
