### 2018-05-24 atmoschem/EmissV#11 @nuest
My apologies for being unclear, I think it is even better to _not_ repeat legal terms, but just say _"This project is published under the X license."_

In you case

```
EmissV is published under the terms of the MIT License. Copyright (c) 2018 Daniel Schuch.
```
