### 2018-05-24 cb4778f
updates for low resulution input
