### 2018-05-24 f98ad16
Just integer are alowed for number of vehicles!
