+++ README.md
@@ -1,54 +1,22 @@
 # EmissV
 [![Travis-CI Build Status](https://travis-ci.org/atmoschem/EmissV.svg?branch=master)](https://travis-ci.org/atmoschem/EmissV) [![Build status](https://ci.appveyor.com/api/projects/status/guuaaklaw6uyn4lj?svg=true)](https://ci.appveyor.com/project/Schuch666/emissv)
 
-## Top-down methods to create vehicular emissions.
+This package provide some methods to create emissions (with a focus on vehicular emissions) for use in numeric air quality models sush as [WRF-Chem](https://ruc.noaa.gov/wrf/wrf-chem/).
 
-Methods to create vehicular and other emissions by a top-down approach for use in numeric air quality models like [WRF-Chem](https://ruc.noaa.gov/wrf/wrf-chem/).
+## To install
 
-![Emissions using nocturnal lights](https://raw.githubusercontent.com/atmoschem/EmissV/master/example.jpg)
+```r
+# install.packages("devtools")
+devtools::install_github("atmoschem/EmissV")
+library(EmissV)
+```
 
-functions:
+### System dependencies 
 
-- read: read global inventories in netcdf format
-- vehicles: tool to set-up vehicle data.table
-- emissionFactor: tool to set-up emission factors data.table
-- gridInfo: read grid information from a NetCDF file
-- pointSource: emissions from point sources
-- plumeRise: calculate plume rise (in progress)
-- rasterSource: distribution of emissions by a georeferenced image
-- lineSource: distribution of emissions by line vectors
-- streetDist: distribution by OpenStreetnMap street (in progress)
-- areaSource: distribution of emissions by region
-- totalEmission: total emissions
-- totalVOC: total VOCs emission
-- emission: Emissions to atmospheric models
+EmissV import functions from [ncdf4](http://cran.r-project.org/package=ncdf4) for read model information, [raster](http://cran.r-project.org/package=raster) to process grinded information and [units](https://github.com/edzer/units/).
 
-datasets:
-
-- Perfil: vehicle counting profile for vehicular activity
-- Sample of an image of persistent lights of the Defense Meteorological Satellite Program (DMSP)
-- CETESB 2015 emission factors as ```emissionFactor(example=T)```
-- DETRAN 2016 data and SP vahicle distribution as ```vehicles(example=T)```
-- Shapefiles for Brasil states
-
-## Packages needed
-
-EmissV uses functions from this packages and will install automatically.
-
-- [raster](http://cran.r-project.org/package=raster)
-- [sp](https://github.com/edzer/sp/)
-- [ncdf4](http://cran.r-project.org/package=ncdf4)
-- [units](https://github.com/edzer/units/)
-- [data.table](https://cran.r-project.org/package=data.table)
-- [sf](https://github.com/r-spatial/sf)
-- [spatstat](https://cran.r-project.org/package=spatstat)
-- [maptools](https://cran.r-project.org/package=maptools)
-
-The dependency of the packages [spatsts](https://cran.r-project.org/package=spatstat), [maptools](https://cran.r-project.org/package=maptools) and [sp](https://github.com/edzer/sp/) will be removed son.
-
-## Libraries for Ubuntu
-The following libraries are required for installation on ubuntu:
-
+### Ubuntu instalation
+The following steps are required for installation on Ubuntu:
 ```bash
   sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
   sudo apt-get --yes --force-yes update -qq
@@ -60,26 +28,140 @@ The following libraries are required for installation on ubuntu:
[…]
