+++ R/gridinfo.R
@@ -46,7 +46,7 @@ gridInfo <- function(file = file.choose(),z=F,verbose = T){
       HGT <- ncdf4::ncvar_get(wrf,varid = "HGT") # 2d
       z   <- PH
       for(i in 1:dim(PH)[3]){
-        z[,,i]   <- (PH[,,i] + PHB[,,i])/9.8 - HGT # 9.81 retornava valores negativos ~10-5
+        z[,,i]   <- (PH[,,i] + PHB[,,i])/9.8 - HGT # 9.81 return values < 0, ~10-5
       }
     }else{
       z <- NA
