+++ paper.md
@@ -25,19 +25,19 @@ affiliations:
 
 # Summary
 
-Air quality models need input data containing information about atmosphere (like temperature, wind, humidity), terrestrial data (like terrain, landuse, soil types) and emissions. Therefore, the emission inventories are easily seen as the scapegoat if a mismatch is found between modelled and observed concentrations of air pollutants [@PullesHeslinga2010].
+Air quality models need input data containing information about atmosphere (like temperature, wind, humidity), terrestrial data (like terrain, landuse, soil types) and emissions. Therefore, the emission inventories are easily seen as the scapegoat if a mismatch is found between modelled and observed concentrations of air pollutants [@PullesHeslinga2010]. The anthropogenic emissions, especially vehicular emissions, are highly dependent on human activity and are constantly changing due to factors ranging from economic (such as the state of conservation of the fleet, renewal of the fleet and the price of fuel) to legal aspects (such as the vehicle routing).
 
-The **EmissV** is an R package that estimates vehicular emissions by a top-down approach. The following steps show an example workflow for calculating vehicular emissions, these emissions are initially temporally and spatially disaggregated, and then distributed spatially and temporally.
+The **EmissV** is an R package that estimates vehicular emissions by a top-down approach, the emissions are calculated using the statistical description of the fleet at avaliable level (National, Estadual, City, etc). The following steps show an example workflow for calculating vehicular emissions, these emissions are initially temporally and spatially disaggregated, and then distributed spatially and temporally.
 
-**I.** Total: emission of pollutants is estimated from the fleet, use and emission factors and for each region.
+**I.** Total: emission of pollutants is estimated from the fleet, use and emission factors and for the interest area (cities, states, countries, etc).
 
 **II.** Spatial distribution: The package has functions to read information from tables, georeferenced images (tiff), shapefiles (sh), OpenStreet maps (osm), global inventories in NetCDF format (nc) to calculate point, line and area sources.
 
-**III.** Emission calculation: calculate the final emission from all different sources and converts to model unit and resolution.
+**III.** Emission calculation: calculate the final emission from all different sources and converts to model units and resolution.
 
 **IV.** Temporal distribution: the package has a set of hourly profiles that represent the mean activity for each day of the week calculated from traffic counts of toll stations located in São Paulo city.
 
-The package also has additional functions for extract information directly from WRF-Chem files and to estimate the emissions emitted in the form of exhaust (exhaust), liquid (sump and evaporative) and vapors (fuel transfer operations) of volatile organic compounds.
+The package has additional functions for read netcdf data, create line and point sources (with plume rise) and to estimate the total emissions of of volatile organic compounds from exhaust (through the exhaust pipe), liquid (carter and evaporative) and vapor (fuel transfer operations).
 
 ## Functions and data
 
[…]
