+++ paper.bib
@@ -1,8 +1,25 @@
+@Manual{ncdftools,
+  title = {ncdf.tools: Easier 'NetCDF' File Handling},
+  author = {Jannis v. Buttlar},
+  year = {2015},
+  note = {R package version 0.7.1.295},
+  url = {https://CRAN.R-project.org/package=ncdf.tools},
+}
+
+@Manual{RNetCDF,
+  title = {RNetCDF: Interface to NetCDF Datasets},
+  author = {Pavel Michna and with contributions from Milton Woods},
+  year = {2017},
+  note = {R package version 1.9-1},
+  url = {https://CRAN.R-project.org/package=RNetCDF},
+}
+
 @article{PullesHeslinga2010,
   title={The art of emission inventorying},
   author={Pulles, Tim and Heslinga, Dick},
   journal={TNO, Utrecht},
-  year={2010}
+  year={2010},
+  doi = {doi.org/10.13140/RG.2.1.2082.8007},
 }
 
 @book{cetesbEV2015,
@@ -29,6 +46,7 @@ Files},
     year = {2018},
     note = {R package version 0.3.0},
     url = {https://github.com/atmoschem/eixport},
+    doi = {10.21105/joss.00607},
   }
 
 @article{pulles2010art,
