+++ R/totalvoc.R
@@ -20,7 +20,6 @@
 #'@import units
 #'
 #'@examples
-#' # Do not run
 #'
 #' veic <- vehicles(example = TRUE)
 #'
