+++ man/pointSource.Rd
@@ -20,7 +20,6 @@ a raster
 Transform a set of points into a grinded output
 }
 \examples{
-# Do not run
 d1 <- gridInfo(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d01",sep=""))
 
 p = data.frame(lat      = c(-22,-22,-23.5),
