+++ man/perfil.Rd
@@ -34,11 +34,12 @@ plot.perfil <- function(per = perfil$LDV, text="", color = "#0000FFBB"){
   axis(2)
   box()
 }
-
+\dontrun{
 # view all profiles in perfil data
 for(i in 1:length(names(perfil))){
   print(paste("profile",i,names(perfil)[i]))
   plot.perfil(perfil[[i]],names(perfil)[i])
+}
 }
 }
 }

[…]
