### 2018-05-26 #662 @Schuch666
Hi @nuest and @benmarwick, thanks for the suggestions, some corrections have already been made to the package and the suggestions in the article have already been incorporated.

@nuest I make some changes in the paper (and in the repository)

- added to the paper a reason to emission inventories being the scapegoat of air quality model input
- The package was tested for some cases in São Paulo and some Brazil adjacent states, for now, the main reason is that São Paulo has a Pollution problem and an available air quality monitory network to compare the results
- I changed the abbreviation for states names
- that's right, the file `sample.tiff` was changed for `dmsp.tiff`, this is a smaller file (for CRAN submission).
- The differences of area source are mainly because of the difference of resolution of the image and the loss of accuracy, it is not a problem (but is an updated version).
- The figure 1 shows `g/d` on each image `pixel`.
- The second figure is created in the `emission` function with `plot = T`, the reason is that output is a matrix with can be written in model input files (with the cited packages) and is not georeferenced
- About the error: the function `rasterSource` was updated (to always deal with `NAs`)
- The References was added to paper, I cited the packages that can be written in netcdf format files, `eixport` has specific functions for write in wrf input files

@benmarwick  the main problem is the need of input data for air quality models, depending on the region there is no information about the temporal or spatial distribution of the vehicular emissions or in some cases, the information is at a national level and not updated for decades. The audience is students and researchers who are interesting in using numerical air quality models but don't have a good inventory or wants an alternative inventory.

The function `vehicles` have more detailed description and the number of vehicles is integer now.

I'm Working on the package remain Issues!
