### 2018-05-26 atmoschem/EmissV#12 @Schuch666
Hi @nuest

I updated manually the `README.md` file:
- Added windows installation notes
- Needed packages list (added information and removed some packages)
- Added Contributor code of conduct / Contributing / License
- Link for `udunits` and `sf` installation
- `yum install` updated

I'm reading the `usethis` documentation, is interesting, I will use this package in the future for sure
