### 2018-05-26 atmoschem/EmissV#14 @Schuch666
@nuest @benmarwick

some package examples was updated, accepting new name suggestions for the examples objects names :)

package comments updated, the `plumeRise` was a case of a new function in test phase
