### 2018-05-26 atmoschem/EmissV#13 @Schuch666
Hi @nuest , that's right, Travis and Appveyor only test the installation

I am learning how to implement the `codecov` on `EmissV`, [this](https://walczak.org/2017/06/how-to-add-code-coverage-codecov-to-your-r-package) is a good way to do this? I already tried to do this without success.

The examples was updated, now the use of `\dontrun{}` codes are limited to 3 cases:
- The function takes much time (`streetDist` and `lineSource`)
- Plot (no need to test the visualization)
- Functions that require additional files like osm or NetCDF (`streetDist` and `read`)
