### 2018-05-28 a945062
update for CRAN
