+++ R/pointSource.R
@@ -21,7 +21,7 @@
 #'                emission = c(666,444,111  ) )
 #'
 #' p_emissions <- pointSource(emissions = p, grid = d1)
-#' \dontrun{
+#' \donttest{
 #' sp::spplot(p_emissions,scales = list(draw=TRUE), ylab="Lat", xlab="Lon",
 #'            main = "3 point sources for domain d1")
 #'}
