+++ R/read.R
@@ -19,7 +19,7 @@
 #'
 #' @source read abbout EDGAR at http://edgar.jrc.ec.europa.eu
 #'
-#' @examples \dontrun{
+#' @examples \donttest{
 #' # Do not run
 #' d1     <- gridInfo(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d01",sep=""))
 #' d2     <- gridInfo(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep=""))
