+++ R/perfil.R
@@ -11,7 +11,7 @@
 #'
 #' # load the data
 #' data(perfil)
-#' \dontrun{
+#' \donttest{
 #' # function to simple view
 #' plot.perfil <- function(per = perfil$LDV, text="", color = "#0000FFBB"){
 #'   plot(per[,1],ty = "l", ylim = range(per),axe = FALSE,
@@ -26,7 +26,7 @@
 #'   axis(2)
 #'   box()
 #' }
-#'\dontrun{
+#'\donttest{
 #' # view all profiles in perfil data
 #' for(i in 1:length(names(perfil))){
 #'   print(paste("profile",i,names(perfil)[i]))
