+++ R/gridinfo.R
@@ -13,7 +13,6 @@
 #' @export
 #'
 #' @examples
-#' # Do not run
 #' grid_d1 <- gridInfo(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d01",sep=""))
 #' \donttest{
 #' grid_d2 <- gridInfo(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep=""))
