+++ R/lineSource.R
@@ -6,11 +6,13 @@
 #' @param grid grid object with the grid information
 #' @param as_raster output format, TRUE for raster, FALSE for matrix
 #' @param verbose display additional information
-#'
+#' @param fest fun (or not)
 #' @export
 #'
 #' @importFrom methods as
 #' @importFrom spatstat owin pixellate.psp as.psp
+#' @importFrom vein emis_grid
+#' @importFrom eixport wrf_grid
 #' @import maptools raster
 #'
 #' @seealso \code{\link{gridInfo}} and \code{\link{rasterSource}}
@@ -21,13 +23,16 @@
 #'                         source = osmar::osmsource_file(paste(system.file("extdata",
 #'                         package="EmissV"),"/streets.osm.xz",sep="")))
 #' roads <- osmar::as_sp(roads,what = "lines")
+#' #roads <- sf::st_as_sf(roads)
 #'
 #' d3    <- gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d03"))
 #'
 #' roadLength <- lineSource(roads,d3,as_raster=TRUE)
 #' sp::spplot(roadLength, scales = list(draw=TRUE), ylab="Lat", xlab="Lon",main="Length of roads",
 #'            sp.layout=list("sp.lines", roads))
+#' roadLengthraster <- lineSource(roads,d3,as_raster=TRUE, fest = TRUE)
 #' }
+#'
 #'
 #'@source OpenstreetMap data avaliable \url{https://www.openstreetmap.org/}
 #'
@@ -36,15 +41,42 @@
 # https://gis.stackexchange.com/questions/119993/convert-line-shapefile-to-raster-value-total-length-of-lines-within-cell
 
 
-lineSource <- function(s,grid,as_raster = F,verbose = T){
+lineSource <- function(s, grid, as_raster = F,verbose = T, fest = F){
 
   print("take a coffee, this function may take a few minutes ...")
+  if (!fest) {
+   g <- eixport::wrf_grid(filewrf = grid$File,
+                           type = "wrfinput",
+                           matrix = FALSE,
+                           epsg = 4326)
+    roads2 <- s
+    #  Calcula comprimento
+    roads2$length <- sf::st_length(sf::st_as_sf(roads))
+    # deixa so o campo length
+    roads3 <- roads2[, "length"]
+    # calcula comprimento de rua em cada celula de grade
+    roads4 <- vein::emis_grid(spobj = roads3, g = g, sr = 4326, type = "lines")
+    # normaliza
+    roads4$length <- roads4$length / sum(roads4$length)
+    # converte Spatial
+    roads4sp <- as(roads4, "Spatial")
+    # rasteriza
+    r <- raster::raster(ncol = grid$Horizontal[1], nrow = grid$Horizontal[2])
+    raster::extent(r) <- raster::extent(roads4sp)
+    r <- raster::rasterize(roads4sp, r, field = roads4sp$length,
+                           update = TRUE, updateValue = "NA")
+    if(as_raster){
+      return(r)
+    } else {
+      return(raster::as.matrix(roadLength))
+    }
+  } else {
 
   if(verbose) print("cropping data for domain (1 of 4) ...")
   x   <- grid$Box$x
   y   <- grid$Box$y
   box <- sp::bbox(sp::SpatialPoints(cbind(x,y)))
-  s   <- raster::crop(s,box)
+  s   <- raster::crop(s,box) # tira ruas fora do dominio
 
   if(verbose) print("converting to a line segment pattern object (2 of 4) ...")
   roadsPSP <- spatstat::as.psp(as(s, 'SpatialLines'))
[…]
