+++ NAMESPACE
@@ -19,6 +19,7 @@ import(raster)
 import(sp)
 import(units)
 importFrom(data.table,data.table)
+importFrom(eixport,wrf_grid)
 importFrom(methods,as)
 importFrom(sf,st_cast)
 importFrom(sf,st_dimension)
@@ -31,3 +32,4 @@ importFrom(spatstat,owin)
 importFrom(spatstat,pixellate.psp)
 importFrom(units,as_units)
 importFrom(units,set_units)
+importFrom(vein,emis_grid)
