### 2018-06-02 b8ec34f
CRAN VERSION!!!!!!
