+++ NAMESPACE
@@ -9,7 +9,6 @@ export(plumeRise)
 export(pointSource)
 export(rasterSource)
 export(read)
-export(streetDist)
 export(totalEmission)
 export(totalVOC)
 export(vehicles)
@@ -18,14 +17,7 @@ import(ncdf4)
 import(raster)
 import(sp)
 import(units)
-importFrom(data.table,data.table)
 importFrom(methods,as)
-importFrom(sf,st_cast)
-importFrom(sf,st_dimension)
-importFrom(sf,st_intersection)
-importFrom(sf,st_length)
-importFrom(sf,st_sf)
-importFrom(sf,st_transform)
 importFrom(spatstat,as.psp)
 importFrom(spatstat,owin)
 importFrom(spatstat,pixellate.psp)
