### 2018-06-04 d9c0f90
Update paper.md
