### 2018-06-04 47fecf5
Update paper.md
