+++ paper.md
@@ -29,15 +29,15 @@ Air quality models need input data containing information about atmosphere (sush
 
 The **EmissV** is an R package that estimates vehicular emissions by a top-down approach, the emissions are calculated using the statistical description of the fleet at avaliable level (National, Estadual, City, etc). The following steps show an example workflow for calculating vehicular emissions, these emissions are initially temporally and spatially disaggregated, and then distributed spatially and temporally.
 
-**I.** Total: emission of pollutants is estimated from the fleet, use and emission factors for each interest area (cities, states, countries, etc).
+**I.** Total: emission of pollutants is estimated from the fleet (number of vehicles), use (km/day) and emission factors (g/km) for each interest area (cities, states, countries, etc) or alternatively the totals of some inventory can be used.
 
-**II.** Spatial distribution: The package has functions to read information from tables, georeferenced images (tiff), shapefiles (sh), openStreetmap data (osm), global inventories in NetCDF format (nc) to calculate point, line and area sources.
+**II.** Spatial distribution: the package has functions to read information from tables, georeferenced images (tiff), shapefiles (sh), openstreetmap data (osm), global inventories in NetCDF format (nc) to calculate point, line and area sources.
 
 **III.** Emission calculation: calculate the final emission from all different sources and converts to model units and resolution.
 
-**IV.** Temporal distribution: the package has a set of hourly profiles that represent the mean activity for each day of the week calculated from traffic counts of toll stations located in São Paulo city.
+**IV.** Temporal distribution: a set of hourly profiles that represent the mean activity (by hour and day of the week) calculated from traffic counts of toll stations located at São Paulo city area available for apply in the emissions.
 
-The package has additional functions for read netcdf data, create line and point sources (with plume rise) and to estimate the total emissions of of volatile organic compounds from exhaust (through the exhaust pipe), liquid (carter and evaporative) and vapor (fuel transfer operations).
+The package has additional functions for create emissions from individual sources (counting with some plume rise parameterizations) and to estimate the vehicular emissions of volatile organic compounds from exhaust (through the exhaust pipe), liquid (carter and evaporative) and vapor (fuel transfer operations).
 
 ## Functions and data
 
@@ -62,7 +62,7 @@ The package has additional functions for read netcdf data, create line and point
 
 ## Examples
 
-The following example creates an area source for São Paulo State (Brasil). The `vehicles` function creates a data.frame with information about the São Paulo Fleet, the `emissionFactors` create a a data.frame with emission factors for CO and PM [@cetesbEV2015]. The `totalEmission` calculate the total of CO for these vehicles and this emission factors. The next 3 lines open different data: a shapefile, a raster and read a wrf file all this data are the input for `areaSouce` that creates an area source based in an image of persistent lights of the Defense Meteorological Satellite Program (DMSP) for São Paulo and Minas Gerais (two states of Brasil) and finally the function `emission` calculate the CO emissions.
[…]
