+++ paper.bib
@@ -93,14 +93,6 @@ url = {https://www.geosci-model-dev-discuss.net/gmd-2017-193/},
 doi = {10.5194/gmd-2017-193}
 }
 
-@Manual{emissv,
-    title = {EmissV: Top-down methods to create vehicular emissions},
-    author = {Daniel Schuch},
-    year = {2017},
-    note = {R package version 0.664.1},
-    url = {https://github.com/atmoschem/EmissV}
-}
-
 @article{freitas2011prep,
   title={PREP-CHEM-SRC-1.0: a preprocessor of trace gas and aerosol emission fields for regional and global atmospheric chemistry models},
   author={Freitas, SR and Longo, KM and Alonso, MF and Pirre, M and Marecal, V and Grell, G and Stockler, R and Mello, RF and G{\'a}cita, M S{\'a}nchez},
