+++ paper.bib
@@ -1,3 +1,10 @@
+@Manual{detran2016,
+  title = {Relatórios Estatísticos: Frota de Veículos de 2015},
+  author = {DETRAN},
+  year = {2015},
+  url = {https://www.denatran.gov.br/estatistica/257-frota-2015},
+}
+
 @Manual{ncdftools,
   title = {ncdf.tools: Easier 'NetCDF' File Handling},
   author = {Jannis v. Buttlar},
