### 2018-06-04 7adc661
Update paper.md
