### 2018-06-04 7a3673a
Update paper.md
