+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'EmissV: An R package to create vehicular and other emissions for air quality models'
+title: 'EmissV: A R package to create vehicular and other emissions for air quality models'
 authors:
 - affiliation: '1'
   name: Daniel Schuch
@@ -25,7 +25,7 @@ affiliations:
 
 # Summary
 
-Air quality models need input data containing information about atmosphere (sush as temperature, wind, humidity), terrestrial data (sush as terrain, landuse, soil types) and emissions. Therefore, the emission inventories are easily seen as the scapegoat if a mismatch is found between modelled and observed concentrations of air pollutants [@PullesHeslinga2010]. The anthropogenic emissions, especially vehicular emissions, are highly dependent on human activity and constantly changing due to various factors ranging from economic (such as the state of conservation of the fleet, renewal of the fleet and the price of fuel) to legal aspects (such as the vehicle routing).
+Air quality models need input data containing information about atmosphere (such as temperature, wind, humidity), terrestrial data (such as terrain, landuse, soil types) and emissions. Therefore, the emission inventories are easily seen as the scapegoat if a mismatch is found between modelled and observed concentrations of air pollutants [@PullesHeslinga2010]. The anthropogenic emissions, especially vehicular emissions, are highly dependent on human activity and constantly changing due to various factors ranging from economic (such as the state of conservation of the fleet, renewal of the fleet and the price of fuel) to legal aspects (such as the vehicle routing).
 
 The **EmissV** is an R package that estimates vehicular emissions by a top-down approach, the emissions are calculated using the statistical description of the fleet at avaliable level (National, Estadual, City, etc). The following steps show an example workflow for calculating vehicular emissions, these emissions are initially temporally and spatially disaggregated, and then distributed spatially and temporally.
 
