### 2018-06-04 52bb2b7
Update paper.md
