### 2018-06-04 3a1d580
Update paper.md
