### 2018-06-04 e4335c1
Update paper.md
