### 2018-06-04 905c318
Update paper.md
