### 2018-06-04 a62dab1
Update emission.R
