### 2018-06-04 839a1b5
Update paper.md
