### 2018-06-04 6e7a395
Update paper.md
