### 2018-06-05 f08c25d
Update paper.md
