+++ man/emissionFactor.Rd
@@ -30,10 +30,10 @@ EF <- emissionFactor(example = TRUE)
 # or the code for the same result
 EF <- emissionFactor(ef = list(CO = c(1.75,10.04,0.39,0.45,0.77,1.48,1.61,0.75),
                                PM = c(0.0013,0.0,0.0010,0.0612,0.1052,0.1693,0.0,0.0)),
-                     vnames = c("Light duty Vehicles Gasohol","Light Duty Vehicles Ethanol",
-                                "Light Duty Vehicles Flex","Diesel trucks","Diesel urban busses",
-                                "Diesel intercity busses","Gasohol motorcycles",
-                                "Flex motorcycles"))
+                     vnames = c("Light Duty Vehicles Gasohol","Light Duty Vehicles Ethanol",
+                                "Light Duty Vehicles Flex","Diesel Trucks","Diesel Urban Busses",
+                                "Diesel Intercity Busses","Gasohol Motorcycles",
+                                "Flex Motorcycles"))
 
 }
 \seealso{
