### 2018-06-05 #662 @Schuch666
Hi @nuest , @benmarwick and @arfon

I have some news: EmissV was accepted on [CRAN](https://cran.r-project.org/web/packages/EmissV/).

The paper was revised:
 - added more detailed description of the problem and the package
 - the example was slightly reformulated (the dmsp_hi-res.tiff was added to the repository and example)
 - added references
 - grammatical and orthographic corrections

When you have time, take a look.

thank you

PS: We are planning the use of EmissV in some Argentine cities in the next months.
