### 2018-06-11 #662 @arfon
> I have some news: EmissV was accepted on CRAN.

Great! Thanks for the update @Schuch666.

@nuest, @benmarwick - could you come and take another look at this sometime in the next couple of weeks?
