+++ R/lineSource.R
@@ -6,14 +6,13 @@
 #' @param grid grid object with the grid information
 #' @param as_raster output format, TRUE for raster, FALSE for matrix
 #' @param verbose display additional information
-#' @param fast fun (or not)
 #' @export
 #'
 #' @importFrom methods as
-#' @importFrom spatstat owin pixellate.psp as.psp
-#' @importFrom vein emis_grid
-#' @importFrom eixport wrf_grid
-#' @import maptools raster
+#' @importFrom data.table data.table
+#' @importFrom data.table .SD
+#' @import sf
+#' @import raster
 #'
 #' @seealso \code{\link{gridInfo}} and \code{\link{rasterSource}}
 #'
@@ -22,40 +21,161 @@
 #' roads <- osmar::get_osm(osmar::complete_file(),
 #'                         source = osmar::osmsource_file(paste(system.file("extdata",
 #'                         package="EmissV"),"/streets.osm.xz",sep="")))
-#' roads <- osmar::as_sp(roads,what = "lines")
-#' #roads <- sf::st_as_sf(roads)
+#' road_lines <- osmar::as_sp(roads,what = "lines")
+#' roads <- sf::st_as_sf(road_lines)
 #'
 #' d3    <- gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d03"))
 #'
 #' roadLength <- lineSource(roads,d3,as_raster=TRUE)
 #' sp::spplot(roadLength, scales = list(draw=TRUE), ylab="Lat", xlab="Lon",main="Length of roads",
-#'            sp.layout=list("sp.lines", roads))
-#' roadLengthraster <- lineSource(roads,d3,as_raster=TRUE, fest = TRUE)
+#'            sp.layout=list("sp.lines", road_lines))
 #' }
 #'
 #'
 #'@source OpenstreetMap data avaliable \url{https://www.openstreetmap.org/} and \url{https://download.geofabrik.de/}
 #'
 
-# OLD algorithm source:
-# https://gis.stackexchange.com/questions/119993/convert-line-shapefile-to-raster-value-total-length-of-lines-within-cell
+lineSource <- function(s, grid, as_raster = F,verbose = T){
 
+  wrf_grid <- function(filewrf, type = "wrfinput", matrix = F, epsg = 4326){
+    cat(paste("using grid info from:", filewrf, "\n"))
+    wrf <- ncdf4::nc_open(filewrf)
+    if(type == "wrfinput"){
+      lat    <- ncdf4::ncvar_get(wrf, varid = "XLAT")
+      lon    <- ncdf4::ncvar_get(wrf, varid = "XLONG")
+    } else if(type == "geo"){
+      lat    <- ncdf4::ncvar_get(wrf, varid = "XLAT_M")
+      lon    <- ncdf4::ncvar_get(wrf, varid = "XLONG_M")
+    }
+    time   <- ncdf4::ncvar_get(wrf, varid = "Times")
+    dx     <- ncdf4::ncatt_get(wrf, varid = 0,
+                               attname = "DX")$value
+    n.lat  <- ncdf4::ncatt_get(wrf, varid = 0,
+                               attname = "SOUTH-NORTH_PATCH_END_UNSTAG")$value
+    n.lon  <- ncdf4::ncatt_get(wrf, varid = 0,
+                               attname = "WEST-EAST_PATCH_END_UNSTAG")$value
+    cat(paste0("Number of lat points ", n.lat, "\n"))
+    cat(paste0("Number of lon points ", n.lon, "\n"))
+    ncdf4::nc_close(wrf)
+    r.lat  <- range(lon)
+    r.lon  <- range(lat)
+    EM  <- matrix(0, nrow = n.lon, ncol = n.lat)
 
[…]
