### 2018-06-13 4cc4f3e
added automated tests!!!
