+++ NAMESPACE
@@ -13,13 +13,13 @@ export(streetDist)
 export(totalEmission)
 export(totalVOC)
 export(vehicles)
-import(maptools)
 import(ncdf4)
 import(raster)
+import(sf)
 import(sp)
 import(units)
+importFrom(data.table,.SD)
 importFrom(data.table,data.table)
-importFrom(eixport,wrf_grid)
 importFrom(methods,as)
 importFrom(sf,st_cast)
 importFrom(sf,st_dimension)
@@ -27,9 +27,5 @@ importFrom(sf,st_intersection)
 importFrom(sf,st_length)
 importFrom(sf,st_sf)
 importFrom(sf,st_transform)
-importFrom(spatstat,as.psp)
-importFrom(spatstat,owin)
-importFrom(spatstat,pixellate.psp)
 importFrom(units,as_units)
 importFrom(units,set_units)
-importFrom(vein,emis_grid)
