### 2018-06-13 f4670b0
fix for units >= 0.5-1
