+++ R/totalEmission.R
@@ -53,9 +53,8 @@ totalEmission <- function(v,ef,pol,verbose = T){
       }
 
       if(verbose){
-        # units::install_symbolic_unit("y")
-        # y <- units::make_unit("y")
-        units::install_conversion_constant("g/d", "t/year", 365/1000000 )
+        if (utils::packageVersion("units") <= "0.5-1")
+          units::install_conversion_constant("g/d", "t/year", 365/1000000 )
         total_t_y <- units::set_units(total,"t/year")
         print(paste("Total of",pol[i],":",sum(total_t_y),units::deparse_unit(total_t_y)))
       }
