+++ man/totalEmission.Rd
@@ -19,7 +19,7 @@ totalEmission(v, ef, pol, verbose = T)
 \description{
 Caculate the total emission with:
 
-  Emission = sum( Vehicles(n) * Km_day_use(n) * Emission_Factor(n,pollutant) )
+  Emission(pollutant) = sum( Vehicles(n) * Km_day_use(n) * Emission_Factor(n,pollutant) )
 
 where n is the type of the veicle
 }
