+++ tests/testthat.R
@@ -1,3 +1,5 @@
+Sys.setenv("R_TESTS" = "")
+
 library(testthat)
 library(EmissV)
 
