+++ tests/testthat.R
@@ -1,5 +1,3 @@
-Sys.setenv("R_TESTS" = "")
-
 library(testthat)
 library(EmissV)
 
