### 2018-06-14 95de420
Joss paper update
