+++ tests/testthat/test-totalVOC.R
@@ -16,23 +16,20 @@ test_that("multiplication works", {
                                  c("vap_g","vap_e","vap_d","liq_g","liq_e","liq_d","exa_g","exa_e","exa_d")),
                         pol = "xyl")[[1]],
                units::set_units(351179047.072522640228271484375,"MOL/d"))
+
+  expect_equal(totalVOC(vehicles(example = TRUE,verbose = F),
+                        setNames(data.frame(t(matrix(c(0.230,0.00,0.120,0.00,0.00,0.00,0.00,0.00,
+                                                       0.000,0.25,0.120,0.00,0.00,0.00,0.00,0.00,
+                                                       0.000,0.00,0.000,0.00,0.00,0.00,0.00,0.00,
+                                                       2.000,0.00,0.875,0.00,0.00,0.00,1.20,0.00,
+                                                       0.000,1.50,0.875,0.00,0.00,0.00,0.00,1.20,
+                                                       0.000,0.00,0.000,0.00,0.00,0.00,0.00,0.00,
+                                                       0.425,0.00,0.217,0.00,0.00,0.00,1.08,0.00,
+                                                       0.000,1.30,0.217,0.00,0.00,0.00,0.00,1.08,
+                                                       0.000,0.00,0.000,2.05,0.00,0.00,0.00,0.00),
+                                                     ncol = 8, byrow = TRUE)),
+                                            row.names = row.names(vehicles(example = T,verbose = F))),
+                                 c("vap_g","vap_e","vap_d","liq_g","liq_e","liq_d","exa_g","exa_e","exa_d")),
+                        pol = "a")[[1]],
+               units::set_units(NA,"MOL/d"))
 })
-
-
-
-# setNames(data.frame(t(matrix(c(0.230,0.00,0.120,0.00,0.00,0.00,0.00,0.00,
-#                                0.000,0.25,0.120,0.00,0.00,0.00,0.00,0.00,
-#                                0.000,0.00,0.000,0.00,0.00,0.00,0.00,0.00,
-#                                2.000,0.00,0.875,0.00,0.00,0.00,1.20,0.00,
-#                                0.000,1.50,0.875,0.00,0.00,0.00,0.00,1.20,
-#                                0.000,0.00,0.000,0.00,0.00,0.00,0.00,0.00,
-#                                0.425,0.00,0.217,0.00,0.00,0.00,1.08,0.00,
-#                                0.000,1.30,0.217,0.00,0.00,0.00,0.00,1.08,
-#                                0.000,0.00,0.000,2.05,0.00,0.00,0.00,0.00),
-#                            ncol = 8, byrow = TRUE)),
-#                     row.names = row.names(vehicles(example = T,verbose = F))),
-#          c("vap_g","vap_e","vap_d","liq_g","liq_e","liq_d","exa_g","exa_e","exa_d"))
-
-
-
-
