### 2018-06-16 56c8ffd
using cat instead of print
