+++ R/emission.R
@@ -54,8 +54,6 @@ emission <- function(total,pol,area,grid, inventory = NULL,mm = 1, aerosol = F,
     # input is g m-2 s-1
     if(class(inventory)[1]=="RasterLayer"){
       VAR_e <- rasterSource(inventory,grid,verbose = verbose)
-    }else{
-      VAR_e <- inventory
     }
 
     if(aerosol){
