+++ tests/testthat/test-emission.R
@@ -21,6 +21,22 @@ test_that("final emission works", {
   ),
   units::as_units(362.58086806097963972206343896687030792236328125, "ug*m^-2*s^-1"))
 
+  expect_equal(sum(emission(totalEmission(vehicles(example = TRUE,verbose = F),
+                                          emissionFactor(example = TRUE,verbose = F),
+                                          pol = c("CO"),verbose = F),
+                            "CO",
+                            list(SP = areaSource(raster::shapefile(paste0(system.file("extdata", package = "EmissV"),"/BR.shp"))[22,1],
+                                                 raster::raster(paste0(system.file("extdata", package = "EmissV"),"/dmsp.tiff")),
+                                                 gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01")),
+                                                 name = "SP",verbose = F)),
+                            gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01"),verbose = F),
+                            mm=28,
+                            verbose = T,
+                            aerosol = T,
+                            plot = T)
+  ),
+  units::as_units(306.59299639647224466898478567600250244140625, "ug*m^-2*s^-1"))
+
   expect_equal(nrow(emission(inventory = read("edgar_co_test.nc"),pol = "FISH",
                             grid = gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01"),
                                             verbose = F),
