+++ tests/testthat/test-totalEmission.R
@@ -5,4 +5,10 @@ test_that("total emission calcualtions", {
                              emissionFactor(example = T,verbose = F),
                              pol = c("CO","PM"),verbose = F)$CO[[1]],
                units::set_units(2244973629, "g/d"))
+
+  expect_equal(totalEmission(vehicles(example = T,verbose = F),
+                             emissionFactor(example = T,verbose = F),
+                             pol = "FISH",verbose = F)[[1]],
+               units::set_units(0, "g/d"))
+
 })
