+++ R/rasterSource.R
@@ -62,8 +62,11 @@ rasterSource <- function(r,grid,nlevels="all",verbose = T){
     X    <- raster::resample(r,box,method = "bilinear") # non-conservative transformation
     X    <- raster::flip(X,2)
     X    <- raster::t(X)
-    X    <- raster::as.array(X)
-    X    <- X[,,1:nlevels]
+    Y    <- raster::as.matrix(X)
+    X    <- array(NA,c(dim(Y),nlevels))
+    for(i in 1:nlevels){
+      X[,,i] = Y
+    }
     # X    <- X * total_box[1:nlevels]/sum(X) # to conserve mass
 
     if(verbose)
