+++ R/read.R
@@ -15,6 +15,7 @@
 #'
 #' @import raster
 #' @import ncdf4
+#' @import osmar
 #' @importFrom units as_units set_units
 #'
 #' @source read abbout EDGAR at http://edgar.jrc.ec.europa.eu
@@ -33,39 +34,49 @@
 #'}
 
 read <- function(file, version = "EDGAR 4.3.1 v2", as_raster = T, verbose = T){
-  ed   <- ncdf4::nc_open(file[1])
-  name <- names(ed$var)
-  var  <- ncdf4::ncvar_get(ed,name[1])
-  varold <- units::as_units(0.0 * var,"g m-2 s-1")
-  var  <- apply(var,1,rev)
-  r    <- raster::raster(x = 1000 * var,xmn=-180,xmx=180,ymn=-90,ymx=90)
-
-  rz <- raster::raster(0.0 * var,xmn=-180,xmx=180,ymn=-90,ymx=90)
-  values(rz) <- rep(0,ncell(rz))
-  raster::crs(rz) <- "+proj=longlat +ellps=GRS80 +no_defs"
-
-  if(verbose)
-    cat(paste0("reading ",name," (",version,") units are g m-2 s-1 ...\n"))
-
-  for(i in 1:length(file)){
-    cat(paste0("from ",file[i]),"\n")
-    ed   <- ncdf4::nc_open(file[i])
+  if(version == "EDGAR 4.3.1 v2"){
+    ed   <- ncdf4::nc_open(file[1])
     name <- names(ed$var)
     var  <- ncdf4::ncvar_get(ed,name[1])
+    varold <- units::as_units(0.0 * var,"g m-2 s-1")
+    var  <- apply(var,1,rev)
+    r    <- raster::raster(x = 1000 * var,xmn=-180,xmx=180,ymn=-90,ymx=90)
+
+    rz <- raster::raster(0.0 * var,xmn=-180,xmx=180,ymn=-90,ymx=90)
+    values(rz) <- rep(0,ncell(rz))
+    raster::crs(rz) <- "+proj=longlat +ellps=GRS80 +no_defs"
+
+    if(verbose)
+      cat(paste0("reading ",name," (",version,") units are g m-2 s-1 ...\n"))
+
+    for(i in 1:length(file)){
+      cat(paste0("from ",file[i]),"\n")
+      ed   <- ncdf4::nc_open(file[i])
+      name <- names(ed$var)
+      var  <- ncdf4::ncvar_get(ed,name[1])
+      if(as_raster){
+        var <- apply(var,1,rev)
+        r   <- raster::raster(x = 1000 * var,xmn=-180,xmx=180,ymn=-90,ymx=90)
+        raster::crs(r) <- "+proj=longlat +ellps=GRS80 +no_defs"
+        names(r) <- name
+        rz       <- rz + r
+      }else{
+        var    <- units::set_units(1000 * var,"g m-2 s-1")
+        varold <- varold + var
+      }
+    }
     if(as_raster){
-      var <- apply(var,1,rev)
-      r   <- raster::raster(x = 1000 * var,xmn=-180,xmx=180,ymn=-90,ymx=90)
-      raster::crs(r) <- "+proj=longlat +ellps=GRS80 +no_defs"
-      names(r) <- name
-      rz       <- rz + r
+      return(rz)
     }else{
-      var    <- units::set_units(1000 * var,"g m-2 s-1")
-      varold <- varold + var
+      return(var)
     }
   }
-  if(as_raster){
-    return(rz)
-  }else{
-    return(var)
+  if(version == "osm"){
+    cat(paste("reading osm data from",file,"\n"))
+    roads <- suppressWarnings( osmar::get_osm(osmar::complete_file(),
[…]
