+++ .travis.yml
@@ -8,7 +8,7 @@ r:
 r_github_packages: atmoschem/eixport
 
 pandoc: false
-warnings_are_errors: false
+warnings_are_errors: true
 
 before_install:
   - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
@@ -23,7 +23,6 @@ before_install:
 script:
   - R CMD build .
   - R CMD check *tar.gz
-  - cat /home/travis/build/atmoschem/EmissV/EmissV.Rcheck/00check.log
 
 after_success:
 - Rscript -e 'covr::codecov()'
