+++ .travis.yml
@@ -28,4 +28,4 @@ after_success:
 - Rscript -e 'covr::codecov()'
 
 after_failure:
-- Rscript -e 'readLines("/home/travis/build/atmoschem/EmissV/EmissV.Rcheck/00check.log")'
+ - ./travis-tool.sh dump_logs
