+++ NAMESPACE
@@ -14,7 +14,6 @@ export(totalEmission)
 export(totalVOC)
 export(vehicles)
 import(ncdf4)
-import(osmar)
 import(raster)
 import(sf)
 import(sp)
