+++ .travis.yml
@@ -26,6 +26,3 @@ script:
 
 after_success:
 - Rscript -e 'covr::codecov()'
-
-after_failure:
- - ./travis-tool.sh dump_logs
