+++ tests/testthat/test-lineSource.R
@@ -1,8 +1,6 @@
 context("lineSource")
 
 test_that("line emissions works", {
-    library(sf)
-
     expect_equal(raster::cellStats(lineSource(readRDS("sf.Rds"),
                                             gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d03")),
                                             as_raster=TRUE),"sum"),
