+++ R/pointSource.R
@@ -74,12 +74,12 @@ pointSource <- function(emissions, grid, verbose=T){
           cat(paste("  height=", altura[k],"k=",k,"\n"))
         }
       }
-      if(emissions$z[i] < altura[2]){
-        a     <- 1
-        cat(paste("Emission heigh between", altura[1],"and",altura[2],"at k=",a,
-                    "for z=",emissions$z[i]))
-        cat("\n")
-      }
+      # if(emissions$z[i] < altura[2]){
+      #   a     <- 1
+      #   cat(paste("Emission heigh between", altura[1],"and",altura[2],"at k=",a,
+      #               "for z=",emissions$z[i]))
+      #   cat("\n")
+      # }
       for(k in 2:length(altura)){
         if(emissions$z[i] >= altura[k-1] & emissions$z[i] < altura[k]){
           a <- k -1
