+++ tests/testthat/test-streetDist.R
@@ -1,0 +1,8 @@
+context("streetDist")
+
+test_that("multiplication works", {
+  expect_equal(dim(suppressWarnings(  streetDist(emission = 1000000,
+                                                 grid = readRDS("g.Rds"),
+                                                 osm  = readRDS("streets.Rds"))) ),
+               c(6960,3))
+})
