+++ .travis.yml
@@ -8,7 +8,7 @@ r:
 r_github_packages: atmoschem/eixport
 
 pandoc: false
-warnings_are_errors: true
+warnings_are_errors: false
 
 before_install:
   - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
