+++ .travis.yml
@@ -5,8 +5,6 @@ r:
   #- devel
   #- oldrel
 
-r_github_packages: atmoschem/eixport
-
 pandoc: false
 warnings_are_errors: false
 
