### 2018-06-18 de61258
format of meter in messeges
