+++ tests/testthat/test-streetDist.R
@@ -1,13 +1,13 @@
 context("streetDist")
 
 test_that("funcion muy loca", {
-  expect_equal(dim(suppressWarnings(  streetDist(emission = 1000000,epsg =  31983,
-                                                 grid = readRDS("g.Rds"),
-                                                 osm  = readRDS("streets.Rds"))) ),
+  expect_equal(dim(suppressWarnings( streetDist(emission = 1000000,epsg = 31983,
+                                                grid = readRDS("g.Rds"),
+                                                osm  = readRDS("streets.Rds"))) ),
                c(6960,3))
 
-  expect_equal(dim(suppressWarnings(  streetDist(emission = 1000000,
-                                                 grid = readRDS("g.Rds"),
-                                                 osm  = readRDS("streets.Rds"))) ),
+  expect_equal(dim(suppressWarnings( streetDist(emission = 1000000,epsg = NULL,
+                                                grid = readRDS("g.Rds"),
+                                                osm  = readRDS("streets.Rds"))) ),
                c(6960,3))
 })
