+++ tests/testthat/test-emission.R
@@ -1,6 +1,40 @@
 context("emission")
 
 test_that("final emission works", {
+
+  expect_equal(drop_units(emission(totalEmission(vehicles(example = TRUE,verbose = F),
+                                      emissionFactor(example = TRUE,verbose = F),
+                                      pol = c("CO"),verbose = T),
+                        "CO",
+                        list(SP = areaSource(raster::shapefile(paste0(system.file("extdata", package = "EmissV"),"/BR.shp"))[22,1],
+                                             raster::raster(paste0(system.file("extdata", package = "EmissV"),"/dmsp.tiff")),
+                                             gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01")),
+                                             name = "SP",verbose = F),
+                             RJ = areaSource(raster::shapefile(paste0(system.file("extdata", package = "EmissV"),"/BR.shp"))[17,1],
+                                             raster::raster(paste0(system.file("extdata", package = "EmissV"),"/dmsp.tiff")),
+                                             gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01")),
+                                             name = "RJ",verbose = F)),
+                        gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01"),verbose = F),
+                        mm=1,
+                        verbose = T,
+                        aerosol = F))/28,
+               drop_units(emission(totalEmission(vehicles(example = TRUE,verbose = F),
+                                      emissionFactor(example = TRUE,verbose = F),
+                                      pol = c("CO"),verbose = T),
+                        "CO",
+                        list(SP = areaSource(raster::shapefile(paste0(system.file("extdata", package = "EmissV"),"/BR.shp"))[22,1],
+                                             raster::raster(paste0(system.file("extdata", package = "EmissV"),"/dmsp.tiff")),
+                                             gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01")),
+                                             name = "SP",verbose = F),
+                             RJ = areaSource(raster::shapefile(paste0(system.file("extdata", package = "EmissV"),"/BR.shp"))[17,1],
+                                             raster::raster(paste0(system.file("extdata", package = "EmissV"),"/dmsp.tiff")),
+                                             gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01")),
+                                             name = "RJ",verbose = F)),
+                        gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d01"),verbose = F),
+                        mm=28,
+                        verbose = T,
+                        aerosol = F)))
+
   expect_equal(sum(emission(totalEmission(vehicles(example = TRUE,verbose = F),
                                           emissionFactor(example = TRUE,verbose = F),
                                           pol = c("CO"),verbose = T),
