+++ tests/testthat/test-streetDist.R
@@ -6,8 +6,8 @@ test_that("funcion muy loca", {
                                                 osm  = readRDS("streets.Rds"))) ),
                c(6960,3))
 
-  # expect_equal(dim(suppressWarnings( streetDist(emission = 1000000,epsg = NULL,
-  #                                               grid = readRDS("g.Rds"),
-  #                                               osm  = readRDS("streets.Rds"))) ),
-  #              c(6960,3))
+  expect_equal(dim(suppressWarnings( streetDist(emission = 1000000,epsg = NULL,
+                                                grid = readRDS("g.Rds"),
+                                                osm  = readRDS("streets.Rds"))) ),
+               c(6960,3))
 })
