+++ appveyor.yml
@@ -12,7 +12,6 @@ install:
 # Adapt as necessary starting from here
 
 build_script:
-  - travis-tool.sh install_github atmoschem/eixport
   - travis-tool.sh install_deps
 
 test_script:
