+++ tests/testthat/test-lineSource.R
@@ -2,12 +2,14 @@ context("lineSource")
 
 test_that("line emissions works", {
 
-  expect_equal(1,sum(lineSource(readRDS("sf.Rds"),
-                                gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d03")),
-                                as_raster=FALSE)))
+  s <- readRDS("sf.Rds")
+  g <- gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d03"))
 
-  expect_equal(raster::cellStats(lineSource(readRDS("sf.Rds"),
-                                            gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d03")),
-                                            as_raster=TRUE),"sum"),1)
+  oldw <- getOption("warn")
+  options(warn = -1)
 
+  expect_equal(1,sum(lineSource(s,g,as_raster=FALSE)))
+  expect_equal(raster::cellStats(lineSource(s,g,as_raster=TRUE),"sum"),1)
+
+  options(warn = oldw)
 })
