+++ codecov.yml
@@ -3,4 +3,4 @@ language: R
 sudo: false
 cache: packages
 after_success:
-- Rscript -e 'covr::codecov()'
+- Rscript -e 'covr::codecov(quiet = FALSE)'
