+++ .travis.yml
@@ -23,4 +23,4 @@ script:
   - R CMD check *tar.gz
 
 after_success:
-- Rscript -e 'covr::codecov()'
+- Rscript -e 'covr::codecov(quiet = FALSE)'
