+++ tests/testthat/test-lineSource.R
@@ -3,8 +3,11 @@ context("lineSource")
 test_that("line emissions works", {
 
   s <- readRDS("sf.Rds")
-  print(getwd())
   g <- gridInfo(paste0(system.file("extdata", package = "EmissV"),"/wrfinput_d03"))
+
+  library(sf)
+  library(raster)
+  library(ncdf4)
 
   expect_equal(1,sum(lineSource(s,g,as_raster=FALSE)))
   expect_equal(raster::cellStats(lineSource(s,g,as_raster=TRUE),"sum"),1)
