+++ NAMESPACE
@@ -13,6 +13,7 @@ export(streetDist)
 export(totalEmission)
 export(totalVOC)
 export(vehicles)
+import(lwgeom)
 import(ncdf4)
 import(raster)
 import(sf)
