### 2018-06-18 b766a4b
But you learn from what's killing you And this time it's real
