### 2018-06-18 #662 @Schuch666
Automated tests added to the package
