### 2018-06-19 73177bf
doc update / emissv 0.664.7 on cran!
