+++ DESCRIPTION
@@ -27,5 +27,5 @@ License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
-URL: https://atmoschem.github.io/EmissV/index.html
+URL: https://atmoschem.github.io/EmissV
 BugReports: https://github.com/atmoschem/EmissV/issues/
