+++ .travis.yml
@@ -4,6 +4,10 @@ r:
   - release
   #- devel
   #- oldrel
+  
+r_binary_packages:
+  - devtools
+  - roxygen2
 
 pandoc: false
 warnings_are_errors: false
@@ -23,4 +27,5 @@ script:
   - R CMD check *tar.gz
 
 after_success:
-- Rscript -e 'covr::codecov(quiet = FALSE)'
+# - Rscript -e 'covr::codecov(quiet = FALSE)'
+- Rscript -e 'devtools::install();devtools::test()'
