### 2018-07-10 e86b7f6
"sp" and "sf" types in lineSource
