+++ NEWS.md
@@ -1,3 +1,6 @@
+## EmissV 0.664.9 (2018-07-10)
+- fix #20
+
 ## EmissV 0.664.8 (2018-06-19)
 - doc updates
 
