+++ R/streetDist.R
@@ -105,7 +105,7 @@ streetDist <- function(emission = 1,
   osm_te$x <- emission*dist[5]*osm_te$LKM/sum(osm_te$LKM, na.rm = T)
   osm_all <- rbind(osm_m, osm_t, osm_p, osm_s, osm_te)
   if (warnings){
-  osmgrid <- sf::st_intersection(osm_all, grid)
+    osmgrid <- sf::st_intersection(osm_all, grid) # nocov
   } else {
     osmgrid <- suppressWarnings(sf::st_intersection(osm_all, grid))
   }
