### 2018-07-19 f87c78f
paper update
