+++ DESCRIPTION
@@ -29,4 +29,4 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
 URL: https://atmoschem.github.io/EmissV
-BugReports: https://github.com/atmoschem/EmissV/issues/
+BugReports: https://github.com/atmoschem/EmissV/issues
