### 2018-08-02 #662 @ibarraespinosa
hi @arfon, I'm co-author of this paper and package. Is it possible to speed up the review process? Thanks! Sergio
