### 2018-08-07 #662 @nuest
@Schuch666 Thanks for taking the time to work on our suggestions, and for your understanding that this takes longer than it should. Congrats for getting the package on CRAN!
