### 2018-08-07 atmoschem/EmissV#14 @nuest
Why not simply `vehicles_data` ?
