### 2018-08-07 atmoschem/EmissV#13 @nuest
I've looked at some of the test functions, this is a good start.

I am not a fan of code coverage, so I can't help you there.
