### 2018-08-07 #662 @nuest
Ok, so PDF re-generation does not seem to work... @arfon did I miss something here?

- Don't want to be too picky on this, but for me the reason to make emission the "scapegoat" must be connected to the other measured variables somehow. Is it the case that athmosphere is not blamed because these values are easier to measure? Is the landscape not blamed because it is not changing?
- I can now compile a PDF from `paper.md` without any issues - good!
- I get slightly different results in some outputs - is that an issue? See e.g.
```
> TOTAL  <- totalEmission(fleet,EF,pol = c("CO"),verbose = T)
Total of CO : 1128297.0993334 t year-1
```
- regarding the differences in the outputs - would you consider updating the paper with a short statement (or the function docs) that the outputs might differ slightly because of... _X_
- I'd like to restate Ben's point about the _statement of need_. AFAICS the first paragraph is almost there. _Do you want to help to better understand anthropogenic emissions?_ _Do you want to stop the emissions being used as a scapegoat?_ What would a user want to... _"estimates vehicular emissions by a top-down approach"._ This is probably totally clear to you, but the first paragraph can just as well welcome any audience.

@arfon Does the author have to make a new release on GitHub? Not sure how to answer the "matches" checklist item. Thanks!
