### 2018-08-07 #662 @arfon
> Ok, so PDF re-generation does not seem to work... @arfon did I miss something here?

☝️ try `@whedon generate pdf`

> @arfon Does the author have to make a new release on GitHub? Not sure how to answer the "matches" checklist item. Thanks!

Basically this is a question about the version number you see here (` 0.664.3`) and what you find on GitHub.
