+++ man/vehicles.Rd
@@ -47,27 +47,27 @@ total_v and area_name must have the same length.
 distribution, type, category (if used), fuel (if used) and vnames (if used) must have the same length.
 }
 \examples{
-veiculos <- vehicles(example = TRUE)
+fleet <- vehicles(example = TRUE)
 
 # or the code bellow for the same result
 # DETRAN 2016 data for total number of vehicles for 5 Brazilian states (Sao Paulo,
 # Rio de Janeiro, Minas Gerais, Parana and Santa Catarina)
 # vahicle distribution of Sao Paulo
 
-veiculos <- vehicles(total_v = c(27332101, 6377484, 10277988, 7140439, 4772160),
-                     area_name = c("SP", "RJ", "MG", "PR", "SC"),
-                     distribution = c( 0.4253, 0.0320, 0.3602, 0.0260,
-                                       0.0290, 0.0008, 0.1181, 0.0086),
-                     category =  c("LDV_E25","LDV_E100","LDV_F","TRUCKS_B5",
-                                   "CBUS_B5","MBUS_B5","MOTO_E25","MOTO_F"),
-                     type = c("LDV", "LDV", "LDV","TRUCKS",
-                              "BUS","BUS","MOTO", "MOTO"),
-                     fuel = c("E25", "E100", "FLEX","B5",
-                              "B5","B5","E25", "FLEX"),
-                     vnames = c("Light duty Vehicles Gasohol","Light Duty Vehicles Ethanol",
-                                "Light Duty Vehicles Flex","Diesel trucks","Diesel urban busses",
-                                "Diesel intercity busses","Gasohol motorcycles",
-                                "Flex motorcycles"))
+fleet <- vehicles(total_v = c(27332101, 6377484, 10277988, 7140439, 4772160),
+                  area_name = c("SP", "RJ", "MG", "PR", "SC"),
+                  distribution = c( 0.4253, 0.0320, 0.3602, 0.0260,
+                                   0.0290, 0.0008, 0.1181, 0.0086),
+                  category =  c("LDV_E25","LDV_E100","LDV_F","TRUCKS_B5",
+                                "CBUS_B5","MBUS_B5","MOTO_E25","MOTO_F"),
+                  type = c("LDV", "LDV", "LDV","TRUCKS",
+                          "BUS","BUS","MOTO", "MOTO"),
+                  fuel = c("E25", "E100", "FLEX","B5",
+                           "B5","B5","E25", "FLEX"),
+                  vnames = c("Light duty Vehicles Gasohol","Light Duty Vehicles Ethanol",
+                             "Light Duty Vehicles Flex","Diesel trucks","Diesel urban busses",
+                             "Diesel intercity busses","Gasohol motorcycles",
+                             "Flex motorcycles"))
 }
 \seealso{
 \code{\link{areaSource}} and  \code{\link{totalEmission}}
