### 2018-08-07 d74c397
Update paper.md
