### 2018-08-17 4fc190d
fix spplot example
