### 2018-09-10 4389080
value change (for the hi-res image file)
