### 2018-09-18 #662 @nuest
I see a few minor questions and some need for rephrasing, as described in https://github.com/openjournals/joss-reviews/issues/662#issuecomment-411080315, other than that I'm done.
