### 2018-09-26 #662 @arfon
> @Schuch666 Please double check the documentation to mention that output values might differ slightly. If users run your algorithms and compare outcomes (e.g. in a reproducible research set-up, a reviewer runs an article manuscript) they should be aware that small differences can occur and why.

Once you've addressed this comment @Schuch666 I think we're good to accept here.
