### 2018-09-27 01a9e74
JOSS review and CRAN fix
