+++ paper.md
@@ -139,6 +139,8 @@ CO_emissions <- emission(TOTAL,"CO",list(SP = Sao_Paulo, MG = Minas_Gerais),
 
 The emissions of CO calculated in this example can be seen in Figure 1 in `g/d` (by pixel) and the final emissions on Figure  2 in `MOL h-1 km-1` (by model grid cell). This emissions can be written to WRF-Chem emission files using some package that makes the interface with NetCDF format such as **ncdf4** [@ncdf4], **RNetCDF** [@RNetCDF], **ncdf.tools** [@ncdftools] or with the **eixport** [@eixport].
 
+Some output values (also figures generated by EmissV) might differ slightly depending on the EmissV package-version (as well as different versios of ncdf4, units, raster, sf, or lwgeom) and changes to the example-files.
+
 The R package EmissV is available at the repository  https://github.com/atmoschem/EmissV. 
 And this installation is tested automatically on Linux via [TravisCI](https://travis-ci.org/atmoschem/EmissV) and Windows via [Appveyor](https://ci.appveyor.com/project/Schuch666/EmissC) continuous integration systems. Also, EmissV is already on CRAN https://CRAN.R-project.org/package=EmissV.
 
