+++ README.md
@@ -155,7 +155,7 @@ Functions:
 - [streetDist](https://atmoschem.github.io/EmissV/reference/streetDist.html): distribution by OpenStreetnMap street (in progress)
 - [areaSource](https://atmoschem.github.io/EmissV/reference/areaSource.html): distribution of emissions by region
 - [totalEmission](https://atmoschem.github.io/EmissV/reference/totalEmission.html): total emissions
-- [totalVOC](https://atmoschem.github.io/EmissV/reference/totalVOC.html): total VOCs emission
+- [totalVOC](https://atmoschem.github.io/EmissV/reference/totalVOC.html): total VOCs emission (this function will be discontinued in the next versions)
 - [emission](https://atmoschem.github.io/EmissV/reference/emission.html): Emissions to atmospheric models
 
 Sample datasets:
