### 2018-09-27 c71b74e
Update paper.md
JOSS note
