### 2018-09-27 #662 @Schuch666
thank you @arfon

I put a note in the text talking about these differences that can appear in the calculated values and in the figures due to changes between versions of the package (in the code or in the sample files) and also of the version of its dependencies.
