### 2018-09-28 bcc8d7c
description for CRAN
