+++ DESCRIPTION
@@ -27,6 +27,6 @@ Suggests: testthat, covr, osmar, rgdal
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0.9000
 URL: https://atmoschem.github.io/EmissV
 BugReports: https://github.com/atmoschem/EmissV/issues
